+++ src/_meta/assets/src/state_machine.rs
@@ -210,8 +210,8 @@ impl State {
         self.assets.clear();
         self.batches.clear();
         self.chunks.clear();
-        self.next_batch_id = Nat::from(1);
-        self.next_chunk_id = Nat::from(1);
+        self.next_batch_id = Nat::from(1 as u64);
+        self.next_chunk_id = Nat::from(1 as u64);
     }
 
     pub fn is_authorized(&self, principal: &Principal) -> bool {
@@ -259,7 +259,7 @@ impl State {
 
     pub fn create_batch(&mut self, now: u64) -> BatchId {
         let batch_id = self.next_batch_id.clone();
-        self.next_batch_id += 1;
+        self.next_batch_id += 1 as u64;
 
         self.batches.insert(
             batch_id.clone(),
@@ -287,7 +287,7 @@ impl State {
         batch.expires_at = Int::from(now + BATCH_EXPIRY_NANOS);
 
         let chunk_id = self.next_chunk_id.clone();
-        self.next_chunk_id += 1;
+        self.next_chunk_id += 1 as u64;
 
         self.chunks.insert(
             chunk_id.clone(),
